// TGS/ClientApp/src/app/Models/user.ts
export interface USER {
  id: any;
  imgPath: any;
  email: any;
  userName: any;
  phoneNumber: any;
  role: any;
  normalizedUserName: any;
  passwordHash: any;
  securityStamp: any;
  concurrencyStamp: any;
  accessFailedCount: any;
  normalizedEmail: any;
  phoneNumberConfirmed: any;
  twoFactorEnabled: any;
  lockoutEnd: any;
  lockoutEnabled: any;
  emailConfirmed: any;

}